requirements.txt
boto3>=1.28.0
python-dotenv>=1.0.0
mcp>=0.1.0
streamlit>=1.28.0 